WIP: Added spec.py script and spec along with initial generated header
This is intial script to parse the spec and create dict containing unique enums and error messages with links.
Currently have checked-in a rough initial header file that was generated from this framework.
Using it in a single error message to demonstrate potential integration.

Working on how to best use/integrate/maintain this going forward.

Added some initial support for create a database of current validation check details
and then updating the header based on latest spec and database to make sure that
unique_ids don't ever change.